Modify installation steps to make conda instructions clearer
The instructions before were a little confusing to me since it setup a
development environment and immediately got out of it. For those without
much experience in Python and python environments, it would probably be
useful to have some insight into why we're setting up the development
environment and how to get in/out of it. This commit also adds a
.gitignore so that a .ipynbcheckpoint file isn't mistakenly commited